glpi_app/requirements.txt
requests
python-dotenv
unstructured
langchain
chromadb
pydantic
typing-extensions
fastapi
uvicorn
boto3
langchain-ibm
reportlab
